Course catalog shows the recommended versions.
When international students and teachers would visit
https://code.org/courses , they were shown a list of courses hardcoded
to the 2017 versions. This change makes it so the catalog instead
displays the recommended version of the course in the user's language
instead. For example, the 2018 version of Course A has been fully
translated to Italian, so the 2018 version of Course A will be displayed
on the course catalog instead. This way students and teachers can easily
find the latest, recommended content.